### 2019-10-09 #1795 @whedon
[PRE REVIEW]: rf: Receiver function calculation in seismology
**Submitting author:** @trichter (<a href="http://orcid.org/0000-0002-8378-559X">Tom Eulenfeld</a>)
**Repository:** <a href="https://github.com/trichter/rf" target ="_blank">https://github.com/trichter/rf</a>
**Version:** v0.8.0
