### 2019-10-09 #1795 @labarba
👋 @katyhuff — Could you handle this submission as editor? It looks like a signal processing application...
**rf: Receiver function calculation in seismology**
https://github.com/trichter/rf
