### 2019-10-10 #1795 @katyhuff
Hi @labarba Thanks for thinking of me, but this is quite far outside my wheelhouse. Perhaps someone in geosciences, like @leouieda or @lheagy, would be much better ?
