### 2019-10-11 #1795 @leouieda
:wave: @ThomasLecocq @seisman would you like to review this submission for JOSS? "rf: Receiver function calculation in seismology" by @trichter (Tom Eulenfeld). Here are links to the [submitted paper](https://github.com/openjournals/joss-papers/blob/joss.01795/joss.01795/10.21105.joss.01795.pdf) and [software repository](https://github.com/trichter/rf). Please see [our documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more information on what a JOSS review is like and feel free to ask any questions you may have (either here or privately by email).
