### 2019-10-14 #1808 @whedon
[REVIEW]: rf: Receiver function calculation in seismology
**Submitting author:** @trichter (<a href="http://orcid.org/0000-0002-8378-559X">Tom Eulenfeld</a>)
**Repository:** <a href="https://github.com/trichter/rf" target ="_blank">https://github.com/trichter/rf</a>
**Version:** 0.9.1
