### 2019-10-17 70bec15
update test_deconvolve_custom_function
