### 2019-10-17 69c72a5
ci: only use toeplitz package for linux
