### 2019-10-17 d01a13f
make toeplitz and tqdm soft dependencies
