### 2019-10-17 89a96d3
ci: update .travis.yml, version 0.9.0
