+++ .travis.yml
@@ -13,7 +13,7 @@ matrix:
     - os: osx
       env: PYTHON_VERSION=3.7
     - os: osx
-      env: PYTHON_VERSION=3.7 USE_OBSPYH5="true"
+      env: PYTHON_VERSION=3.7 USE_OBSPYH5="true" USE_TOEPLITZ="true" NPY_DISTUTILS_APPEND_FLAGS=1
     - os: windows
       env: PYTHON_VERSION=3.7
     - os: windows
@@ -21,8 +21,8 @@ matrix:
 before_install:
   - wget https://raw.githubusercontent.com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh
   - source conda4travis.sh
-  - if [[ "$USE_TOEPLITZ" == "true" ]]; then
-      GF="gfortran_linux-64"; TOEPLITZ="toeplitz";
+  - if [[ $USE_TOEPLITZ == "true" ]]; then
+      GF="gfortran_${TRAVIS_OS_NAME}-64"; TOEPLITZ="toeplitz";
     fi
   - if [[ $USE_OBSPYH5 == "true" ]]; then
       OBSPYH5DEP="h5py"; OBSPYH5="obspyh5";
@@ -39,7 +39,7 @@ before_install:
       $GF $OBSPYH5DEP
       coverage
   - conda activate testenv
-  - if [[ -v USE_OBSPYH5 ]] || [[ -v USE_TOEPLITZ ]]; then
+  - if [[ $USE_TOEPLITZ == "true" ]] || [[ $USE_OBSPYH5 == "true" ]]; then
       pip install $TOEPLITZ $OBSPYH5;
     fi
   - conda list
