### 2019-10-29 trichter/rf#18 @seisman
[JOSS review] Errors when install toeplitz
I'm using macOS. I followed the install instructions (create an environment, and replace `gfortran_linux-64` with `gfortran_osx-64`), and had trouble installing the toeplitz package.

Please see the attached logfile [errors.log](https://github.com/trichter/rf/files/3785345/errors.log).

https://github.com/openjournals/joss-reviews/issues/1808
