### 2019-10-29 trichter/rf#18 @trichter
Does it help if you set  `NPY_DISTUTILS_APPEND_FLAGS` to 1?

```
export NPY_DISTUTILS_APPEND_FLAGS=1
```
