### 2019-10-29 trichter/rf#18 @trichter
Probably, this should be mentioned in the docs. I already set it in my .bashrc.
