### 2019-11-04 trichter/rf#19 @seisman
[JOSS review] Mention the RTZ coordinate system in the method section
https://github.com/trichter/rf/blob/89a96d3fa7c3a9f9f691d6e95131407d630495f5/rf/__init__.py#L25-L28

Both RTZ and LQT coordinate systems are commonly used in receiver function studies. When I read this section, I thought this package only supports LQT coordinate system, then I realized that's not true after reading the API of the [rf](https://rf.readthedocs.io/en/latest/#rf.rfstream.RFStream.rf) function.

It would be better to clarify that this package supports both.

https://github.com/openjournals/joss-reviews/issues/1808
