### 2019-11-04 trichter/rf#20 @seisman
[JOSS review] Add the links of tutorial notebooks to the documentation
There are links of two example notebooks in the README.md, but I can't find them in the documentation.

It would be better to list them in the documentation page, since not all users will read README in this repository.

https://github.com/openjournals/joss-reviews/issues/1808
