### 2019-11-04 trichter/rf#21 @seisman
[JOSS] The second notebook example doesn't work well
https://nbviewer.jupyter.org/github/trichter/notebooks/blob/master/receiver_function_profile_chile.ipynb

Issues:

1. Need to create a "data" directory before running the script
2. I got following errors, possibly an issue of matplotlib or obspy?
```
Traceback (most recent call last):
  File "/Users/seisman/.anaconda/envs/rfenv/lib/python3.7/site-packages/matplotlib/colors.py", line 109, in _sanitize_extrema
    ret = ex.item()
AttributeError: 'str' object has no attribute 'item'

Traceback (most recent call last):
  File "test.py", line 26, in <module>
    inventory.plot(label=False)
  File "/Users/seisman/.anaconda/envs/rfenv/lib/python3.7/site-packages/obspy/core/inventory/inventory.py", line 749, in plot
    title=None, show=False, fig=fig, **kwargs)
  File "/Users/seisman/.anaconda/envs/rfenv/lib/python3.7/site-packages/obspy/imaging/maps.py", line 768, in plot_map
    return plot_cartopy(*args, **kwargs)
  File "/Users/seisman/.anaconda/envs/rfenv/lib/python3.7/site-packages/obspy/imaging/maps.py", line 553, in plot_cartopy
    scal_map = ScalarMappable(norm=Normalize(min_color, max_color),
  File "/Users/seisman/.anaconda/envs/rfenv/lib/python3.7/site-packages/matplotlib/colors.py", line 909, in __init__
    self.vmin = _sanitize_extrema(vmin)
  File "/Users/seisman/.anaconda/envs/rfenv/lib/python3.7/site-packages/matplotlib/colors.py", line 111, in _sanitize_extrema
    ret = float(ex)
ValueError: could not convert string to float: '#b15928'
```

https://github.com/openjournals/joss-reviews/issues/1808
