### 2019-11-04 trichter/rf#19 @seisman
Also need to clarify that in the software paper.
