### 2019-11-05 trichter/rf#21 @trichter
I see, the data directory could be created automatically, if it is not present.

I think the second issue is due to a combination of old obspy and new matplotlib versions. See obspy/obspy#2097 (and others). Can you confirm, that you are not using tha latest ObsPy version `1.1.1`?
