### 2019-11-05 726a117
add note about numpy flag
