### 2019-11-05 c86a7f6
add links of tutorial notebooks to the documentation
