### 2019-11-05 trichter/rf#21 @seisman
You're right, my ObsPy version is `1.1.0`, but I followed the install instruction exactly.
