### 2019-11-05 trichter/rf#21 @trichter
I cannot reproduce your issue on Linux.

But I noticed, that the latest release on conda-forge pinned the numpy version to `<1.16` (conda-forge/obspy-feedstock#26). Your conda solver probably installed a newer numpy version and therefore obspy `1.1.0`. I think the issue could be solved by using the install line

```
conda create -n rfenv obspy tqdm shapely cartopy geographiclib h5py gfortran_osx-64
```

without the dependencies which are itself dependencies of obspy.

If you could confirm this, I 'd be happy to update the documentation.
