### 2019-11-05 trichter/rf#21 @seisman
It works, now I have obspy 1.1.1
