### 2019-11-05 898242b
update install instructions, see #21
