### 2019-11-07 trichter/rf#21 @seisman
Now it works well for me.
