+++ joss/paper.bib
@@ -33,9 +33,13 @@
     url = {http://cds.cern.ch/record/151420}
 }
 
-@misc{telewavesim,
-    title = {Telewavesim: Python package for teleseismic body-wave modeling},
-    author = {Pascal Audet},
+@article{telewavesim,
+    title = {Telewavesim: Python software for teleseismic body wave modeling},
+    author = {Audet, Pascal and Thomson, Colin J. and Bostock, Michael G. and Eulenfeld, Tom},
+    journal = {Journal of Open Source Software},
     year = {2019},
-    doi = {10.5281/zenodo.3387720},
+    volume = {4},
+    number = {44},
+    pages = {1818},
+    doi = {10.21105/joss.01818}
 }
