### 2019-12-06 #1808 @trichter
I updated the reference for the telewavesim package, which was recently published in JOSS.
