### 2019-12-11 360601a
plot_rf can now plot stream with different npts
* see #23
* stack is not plotted in this case
