### 2020-02-28 5c7d2cc
Small additions, windows compatible fortran compiler
