### 2020-02-28 #1808 @ThomasLecocq
All checks above are OK for me, I just proposed some changes to make windows fortran a real thing :)

(I cannot change the checkboxes above, sorry)

and sorry for the awful delay....
