### 2020-04-03 #1808 @leouieda
@trichter I made a few editorial suggestions to the paper. After those are integrated, there a couple of things that you'll need to do before acceptance:

- Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.
- Archive the release on Zenodo/figshare/etc and **post the DOI here**.
