### 2020-04-06 #1808 @trichter
I worked through your posted list:

* Added your editorial suggestions and my orcid to the paper
* Published a new version ``0.9.1``
* Uploaded it to zenodo, the doi is ``10.5281/zenodo.3741366``
