### 2020-04-06 #1808 @leouieda
@trichter thank you for implementing these final fixes. It's my pleasure to recommend your submission for publication in JOSS :tada:

A JOSS Associate Editor-in-Chief will stop by and do the final acceptance and publication (usually quick but please be patient given the current global crisis).

@seisman @ThomasLecocq thank for your reviews of this submission!
