### 2020-04-06 #1808 @danielskatz
@trichter - there are some small changes in https://github.com/trichter/rf/pull/27
