### 2020-04-06 #1808 @trichter
@danielskatz Thanks for these. I merged your branch.
